# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Learning Voice Agent API
  description: |
    AI-powered voice conversation API for learning capture and knowledge management.

    ## Overview
    The Learning Voice Agent provides multiple interfaces for capturing and retrieving
    conversational learning content:

    - **REST API**: Synchronous request/response for simple integrations
    - **WebSocket**: Real-time bidirectional streaming for audio conversations
    - **Twilio Webhooks**: Phone-based voice conversations

    ## Authentication
    Currently, the API does not require authentication for development.
    Production deployments should implement appropriate authentication mechanisms.

    ## Rate Limiting
    No rate limiting is currently enforced. Production deployments should
    implement rate limiting based on session_id or client IP.

    ## Error Handling
    All endpoints return standard HTTP status codes:
    - `200`: Success
    - `400`: Bad Request (invalid input)
    - `403`: Forbidden (invalid Twilio signature)
    - `500`: Internal Server Error
  version: 1.0.0
  contact:
    name: Learning Voice Agent Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (configure as needed)
...
paths:
...
  /twilio/voice:
    post:
      tags:
        - Twilio
      summary: Twilio voice webhook
      description: |
        Main entry point for Twilio voice calls. Handles incoming calls and
        generates TwiML responses for voice interaction.

        This endpoint:
        1. Validates the Twilio request signature (in production)
        2. Creates or retrieves the session based on CallSid
        3. Generates appropriate TwiML based on call status
        4. Updates session metadata asynchronously

        **Note**: This endpoint expects `application/x-www-form-urlencoded` data
        from Twilio, not JSON.
      operationId: handleTwilioVoice
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TwilioVoiceRequest'
      responses:
        '200':
          description: TwiML response generated
          content:
            application/xml:
              schema:
                type: string
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <Response>
                  <Gather input="speech" timeout="3" language="en-US"
                          action="/twilio/process-speech?session_id=twilio_CA123"
                          method="POST" speechTimeout="auto" enhanced="true">
                    <Say voice="Polly.Joanna-Neural" language="en-US">
                      Hello! I'm your learning companion. Tell me what you're working on.
                    </Say>
                  </Gather>
                  <Say>I didn't catch that. Let's try again.</Say>
                  <Redirect>/twilio/voice?session_id=twilio_CA123</Redirect>
                </Response>
        '403':
          description: Invalid Twilio request signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Invalid request signature
  /twilio/process-speech:
    post:
      tags:
        - Twilio
      summary: Process Twilio speech input
      description: |
        Processes transcribed speech from Twilio's speech recognition.
        Generates a conversational response and returns TwiML for continuation.

        The endpoint:
        1. Validates speech confidence (threshold: 0.5)
        2. Retrieves conversation context
        3. Generates AI response using Claude
        4. Detects conversation intent for flow control
        5. Returns appropriate TwiML (continue or end call)
      operationId: processTwilioSpeech
      parameters:
        - name: session_id
          in: query
          required: true
          description: Session identifier from voice webhook
          schema:
            type: string
          example: twilio_CA123456
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                SpeechResult:
                  type: string
                  description: Transcribed speech text
                Confidence:
                  type: number
                  format: float
                  description: Speech recognition confidence (0-1)
      responses:
        '200':
          description: TwiML response with agent reply
          content:
            application/xml:
              schema:
                type: string
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <Response>
                  <Gather input="speech" timeout="5" language="en-US"
                          action="/twilio/process-speech?session_id=twilio_CA123"
                          method="POST" speechTimeout="auto" enhanced="true">
                    <Say voice="Polly.Joanna-Neural" language="en-US">
                      That's interesting! Tell me more about that project.
                    </Say>
                  </Gather>
                  <Say>I didn't catch that. Let's try again.</Say>
                  <Redirect>/twilio/voice?session_id=twilio_CA123</Redirect>
                </Response>
  /twilio/recording:
    post:
      tags:
        - Twilio
      summary: Handle Twilio recording webhook
      description: |
        Receives notification of completed recordings from Twilio.
        Processes the recording asynchronously in the background.

        **Note**: Full recording processing (download, transcribe, respond)
        is implemented as a background task.
      operationId: handleTwilioRecording
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                RecordingUrl:
                  type: string
                  description: URL to download the recording
                CallSid:
                  type: string
                  description: Twilio Call SID
                RecordingSid:
                  type: string
                  description: Twilio Recording SID
                RecordingDuration:
                  type: integer
                  description: Recording duration in seconds
      responses:
        '200':
          description: Recording acknowledged
          content:
            application/xml:
              schema:
                type: string
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <Response>
                  <Say>Got it. Processing your recording.</Say>
                </Response>
components:
  schemas:
...
    TwilioVoiceRequest:
      type: object
      description: Twilio voice webhook request
      properties:
        CallSid:
          type: string
          description: Unique identifier for the call
          example: CA1234567890abcdef
        From:
          type: string
          description: Caller phone number
          example: '+15551234567'
        To:
          type: string
          description: Called phone number
          example: '+15559876543'
        CallStatus:
          type: string
          description: Current call status
          enum:
            - ringing
            - in-progress
            - completed
            - busy
            - failed
            - no-answer
          example: ringing
        RecordingUrl:
          type: string
          description: URL of any recording
          example: https://api.twilio.com/...
        SpeechResult:
          type: string
          description: Transcribed speech (if using Gather)
          example: I want to learn about machine learning
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        detail:
          type: string
          description: Error message
          example: No input provided
...